hotfix 16062023/1
- Fixed issue with os icon not displaying correctly in details
- fixed issue with screenshot not displaying correctly in details
- added progress of operation to all VMs commands
- Added the ability to drag multiple machines
- Added 'preview' icon to the extension manifest
-Added badge to the extension manifest
- Added more OS icons
- added black screen is machine screenshot is not available
- fixed missing group delete icon

+++ src/tree/commands/removeGroup.ts
@@ -10,13 +10,19 @@ export function registerRemoveGroupCommand(context: vscode.ExtensionContext, pro
     vscode.commands.registerCommand(CommandsFlags.treeViewRemoveGroup, async item => {
       const group = Provider.getConfiguration().getVirtualMachineGroup(item.name);
       if (group !== undefined) {
-        const noGroup = Provider.getConfiguration().getVirtualMachineGroup(FLAG_NO_GROUP);
-        group.machines.forEach(vm => {
-          noGroup?.add(vm);
+        const options: string[] = ["Yes", "No"];
+        const confirmation = await vscode.window.showQuickPick(options, {
+          placeHolder: `Are you sure you want to remove group ${item.name}?`
         });
-        Provider.getConfiguration().deleteVirtualMachineGroup(item.name);
-        vscode.commands.executeCommand(CommandsFlags.treeViewRefreshVms);
-        parallelsOutputChannel.appendLine(`Group ${item.name} removed`);
+        if (confirmation !== "Yes") {
+          const noGroup = Provider.getConfiguration().getVirtualMachineGroup(FLAG_NO_GROUP);
+          group.machines.forEach(vm => {
+            noGroup?.add(vm);
+          });
+          Provider.getConfiguration().deleteVirtualMachineGroup(item.name);
+          vscode.commands.executeCommand(CommandsFlags.treeViewRefreshVms);
+          parallelsOutputChannel.appendLine(`Group ${item.name} removed`);
+        }
       }
     })
   );
